pyproject.toml
[project]
name = "ubiwhere"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "django>=5.2.6",
    "djangorestframework>=3.16.1",
    "djangorestframework-gis>=1.2.0",
    "drf-spectacular>=0.28.0",
    "pandas>=2.3.2",
    "psycopg2-binary>=2.9.10",
    "python-dotenv>=1.1.1",
]

[dependency-groups]
dev = [
    "faker>=37.8.0",
    "pytest-cov>=7.0.0",
    "pytest-django>=4.11.1",
]
